Return number of bytes read from dnet_read_data_wait() and callers (C++/Python).

+++ include/elliptics/interface.h
@@ -154,9 +154,11 @@ int dnet_read_object(struct dnet_node *n, struct dnet_io_control *ctl);
  * Read data associated with key @ID and put it to provided @data pointer.
  * Requestted chunk will have at most @size bytes in.
  * It will be read from offset @offset from the start of remote object.
+ *
+ * Returns number of bytes read or negative error code.
  */
-int dnet_read_data_wait(struct dnet_node *n, struct dnet_id *id, void *data,
-		uint64_t offset, uint64_t size);
+ssize_t dnet_read_data_wait(struct dnet_node *n, struct dnet_id *id,
+		void *data, uint64_t offset, uint64_t size);
 
 int dnet_send_read_data(void *state, struct dnet_cmd *cmd, struct dnet_io_attr *io,
 		void *data, int fd, uint64_t offset);
